requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Flask-Migrate
python-dotenv
email-validator
flask-mail
requests
